+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - More space at the right of the up/down buttons used to change effects order.
 In a few systems GTK is configured to always show the scrollbar. In this case
 the down button would be partially hidden.
+- PulseEffects can be reset from the command line "pulseeffects -r".
 
 ### Removed
 - The app menu with the entries quit, about and help was removed. This menu was